Conflicts:
	Gemfile
	Gemfile.lock
	app/assets/javascripts/initializers/current_user.js.coffee
	app/assets/javascripts/models/feedback.js.coffee
	app/assets/javascripts/models/user_flow.js.coffee
	app/assets/javascripts/routes/application_route.js.coffee
	app/assets/javascripts/routes/paper_route.js.coffee
	app/assets/javascripts/templates/journal.hbs
	app/assets/javascripts/templates/journal/flow_manager.hbs
	client/app/controllers/overlays/choose-new-flow-manager-column.coffee
	client/app/controllers/overlays/feedback.coffee
	client/app/controllers/overlays/paper-new.coffee
	client/app/initializers/error-handler.coffee
	client/app/models/card-thumbnail.coffee
	client/app/pods/components/flow-column/component.coffee
	client/app/pods/components/participant-selector/component.coffee
	client/app/pods/flow-manager/route.coffee
	client/app/services/rest-less.coffee

+++ app/models/participation.rb
@@ -18,7 +18,7 @@ class Participation < ActiveRecord::Base
   end
 
   def remove_paper_role
-    if paper.participants.where(id: user.id).none?
+    if user.present? && paper.participants.where(id: user.id).none?
       paper.paper_roles.participants.where(user: user).destroy_all
     end
   end
